moved extendedlogger to within swarming
