+++ nupic/swarming/ExtendedLogger.py
