+++ nupic/swarming/HypersearchWorker.py
@@ -36,7 +36,7 @@ from nupic.data import jsonhelpers
 from nupic.swarming.hypersearch.utils import clippedObj
 from nupic.support import initLogging
 from nupic.support.configuration import Configuration
-from nupic.support.ExtendedLogger import ExtendedLogger
+from nupic.swarming.ExtendedLogger import ExtendedLogger
 from nupic.swarming.hypersearch.errorcodes import ErrorCodes
 from nupic.database.ClientJobsDAO import ClientJobsDAO
 from HypersearchV2 import HypersearchV2
